Fixed warning tests

+++ test/gtc/gtc_integer.cpp
@@ -188,8 +188,8 @@ namespace iround
 
 		for(float f = 0.0f; f < 3.1f; f += 0.05f)
 		{
-			int RoundFast = glm::iround(f);
-			int RoundSTD = glm::round(f);
+			int RoundFast = static_cast<int>(glm::iround(f));
+			int RoundSTD = static_cast<int>(glm::round(f));
 			Error += RoundFast == RoundSTD ? 0 : 1;
 			assert(!Error);
 		}
